Art is everywhere

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 int i;
-long int pf, af, f;
+int pf, af, f;
 
 pf = 1;
 af = 2;
